`ops/numpy.py`: Support `key` as `list` in `GetItem`
When loading a model that contains `GetItem` nodes with multidimensional
indices/slices as `key`, the `key` argument is loaded from JSON as a `list`,
not a `tuple` (because JSON does not have the distinction).

So, treat the `key list` as equivalent to the `key tuple`.
Copying is important: otherwise, the later `pop()` will remove the bound
slice elements from the op itself.

`saving/serialization_lib_test.py`:

* Add `test_numpy_get_item_layer()`:
	test for consistent serialization/deserialization of a model which
	contains `ops.numpy.GetItem`;
	because the test with the existing `assertEqual()` would fail due
	to the implicit mapping of tuples to lists:

* Explicitly handle the implicit mapping of any tuple to list when passing
	through the JSON layer by adding an `assertEqual()` override which
	maps any (nested) `tuple` to `list`

* Consequently add a tuple-case to `test_simple_objects()`